work tested

+++ openwsn-fw/openstack/03b-IPv6/icmpv6rpl.c
@@ -839,7 +839,7 @@ void icmpv6rpl_setDIOPeriod(uint16_t dioPeriod){
    );
 }
 
-void icmpv6rpl_setDAOPeriod(uint32_t daoPeriod){
+void icmpv6rpl_setDAOPeriod(uint16_t daoPeriod){
 //void icmpv6rpl_setDAOPeriod(uint16_t daoPeriod){
    //uint32_t        daoPeriodRandom;
